Move test files from app fabric src/main to src/test.
