+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/DefaultId.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.test.internal;
+package co.cask.cdap.internal;
 
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.proto.Id;
