+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/MockResponder.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package co.cask.cdap.test.internal;
+package co.cask.cdap.internal;
 
 import co.cask.http.AbstractHttpResponder;
 import co.cask.http.ChunkResponder;
