+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/TempFolder.java
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package co.cask.cdap.test.internal;
+package co.cask.cdap.internal;
 
 import com.google.common.base.Throwables;
