+++ cdap-app-fabric/src/test/java/co/cask/cdap/internal/app/queue/SimpleQueueSpecificationGeneratorTest.java
@@ -23,11 +23,11 @@ import co.cask.cdap.app.ApplicationSpecification;
 import co.cask.cdap.app.queue.QueueSpecification;
 import co.cask.cdap.app.queue.QueueSpecificationGenerator;
 import co.cask.cdap.common.conf.Constants;
+import co.cask.cdap.internal.DefaultId;
 import co.cask.cdap.internal.app.ApplicationSpecificationAdapter;
 import co.cask.cdap.internal.app.Specifications;
 import co.cask.cdap.internal.io.ReflectionSchemaGenerator;
 import co.cask.cdap.proto.Id;
-import co.cask.cdap.test.internal.DefaultId;
 import com.google.common.base.Function;
 import com.google.common.base.Joiner;
 import com.google.common.base.Predicate;
