+++ cdap-unit-test/src/main/java/co/cask/cdap/test/ConfigurableTestBase.java
@@ -67,6 +67,7 @@ import co.cask.cdap.explore.executor.ExploreExecutorService;
 import co.cask.cdap.explore.guice.ExploreClientModule;
 import co.cask.cdap.explore.guice.ExploreRuntimeModule;
 import co.cask.cdap.gateway.auth.AuthModule;
+import co.cask.cdap.internal.LocalNamespaceClient;
 import co.cask.cdap.internal.app.namespace.NamespaceAdmin;
 import co.cask.cdap.internal.app.runtime.schedule.SchedulerService;
 import co.cask.cdap.logging.guice.LoggingModules;
@@ -82,7 +83,6 @@ import co.cask.cdap.test.internal.ApplicationManagerFactory;
 import co.cask.cdap.test.internal.DefaultApplicationManager;
 import co.cask.cdap.test.internal.DefaultStreamManager;
 import co.cask.cdap.test.internal.DefaultStreamWriter;
-import co.cask.cdap.test.internal.LocalNamespaceClient;
 import co.cask.cdap.test.internal.LocalStreamWriter;
 import co.cask.cdap.test.internal.StreamManagerFactory;
 import co.cask.cdap.test.internal.StreamWriterFactory;
