+++ cdap-unit-test/src/main/java/co/cask/cdap/test/internal/DefaultAdapterManager.java
@@ -16,6 +16,7 @@
 
 package co.cask.cdap.test.internal;
 
+import co.cask.cdap.internal.AppFabricClient;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.proto.RunRecord;
 import co.cask.cdap.test.AbstractAdapterManager;
