+++ cdap-unit-test/src/main/java/co/cask/cdap/test/internal/DefaultFlowManager.java
@@ -17,6 +17,7 @@
 package co.cask.cdap.test.internal;
 
 import co.cask.cdap.api.metrics.RuntimeMetrics;
+import co.cask.cdap.internal.AppFabricClient;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.test.AbstractProgramManager;
 import co.cask.cdap.test.FlowManager;
