+++ cdap-unit-test/src/main/java/co/cask/cdap/test/internal/DefaultWorkerManager.java
@@ -16,13 +16,12 @@
 
 package co.cask.cdap.test.internal;
 
+import co.cask.cdap.internal.AppFabricClient;
 import co.cask.cdap.proto.Id;
 import co.cask.cdap.test.AbstractProgramManager;
 import co.cask.cdap.test.WorkerManager;
 import com.google.common.base.Preconditions;
 import com.google.common.base.Throwables;
-
-import java.util.concurrent.TimeUnit;
 
 /**
  * A default implementation of {@link WorkerManager}
